Introducing removeme()
Now a user can remove themselves without BP interaction, by calling:
tlosrecovery::removeme(myaccount)

According to TBNOA, we are not permitted to take tokens from accounts which
have been used after launch. So, by calling the function you need to use the
chain, becoming one of the parties to the TBNOA, and therefore immune to the
recovery process.

Also fixed a for loop for *many().